Testcontainers fix
