+++ src/test/java/com/construction_worker_forum_back/integration/chat/MongoTestcontainersConfig.java
@@ -1,0 +1,24 @@
+package com.construction_worker_forum_back.integration.chat;
+
+import org.springframework.boot.autoconfigure.mongo.embedded.EmbeddedMongoAutoConfiguration;
+import org.springframework.boot.test.autoconfigure.data.mongo.DataMongoTest;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.testcontainers.containers.MongoDBContainer;
+import org.testcontainers.junit.jupiter.Testcontainers;
+
+@Testcontainers
+@DataMongoTest(excludeAutoConfiguration = EmbeddedMongoAutoConfiguration.class)
+public abstract class MongoTestcontainersConfig {
+
+    public static MongoDBContainer mongoDBContainer = new MongoDBContainer("mongo");
+
+    static {
+        mongoDBContainer.start();
+    }
+
+    @DynamicPropertySource
+    public static void setProperties(DynamicPropertyRegistry registry) {
+        registry.add("spring.data.mongodb.uri", mongoDBContainer::getReplicaSetUrl);
+    }
+}
